Acceleration for x86 compressed String hashCode with vectorization
For obtaining the hashCode of a compressed String, each iteration of the SSE loop moves 4 bytes from memory to the low doubleword of the XMM register. The stride shift amount is set to 0 for handling bytes. These packed 1 byte integers are then zero extended with instruction pmovzxbd and the same procedure to obtain the hashCode as with decompressed Strings is used.

A vector conversion example has been added to show how this procedure works.

Signed-off-by: Mansoor Saqib <Mansoor.Saqib@ibm.com>